fix: Handle case when user is not found in UserController remove method
The code changes in this commit handle the case when a user is not found in the UserController remove method. If the user is not found, a response object with a failure message is returned. This ensures that the API provides accurate feedback when attempting to remove a user that does not exist.

+++ src/models/user/user.service.ts
@@ -4,6 +4,8 @@ import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { User } from './user.entity';
 
+type TRemoveReturn = { id: string; removed: boolean };
+
 @Injectable()
 export class UserService {
   constructor(
@@ -24,7 +26,12 @@ export class UserService {
     return this.userRepository.findOneBy({ id });
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: string): Promise<TRemoveReturn> {
+    const user = await this.findOne(id);
+
+    if (!user) return null;
+
     await this.userRepository.delete(id);
+    return { id: user.id, removed: true };
   }
 }
